Feature/history

+++ Client/src/core/Window.h
@@ -58,8 +58,8 @@ public:
 #pragma region Other
 
     sf::Vector2f GetCenter() const { return sf::Vector2f(m_Window->getSize().x * 0.5f, m_Window->getSize().y * 0.5f); }
-    int GetWidth() const { return m_Window->getSize().x; }
-    int GetHeight() const { return m_Window->getSize().y; }
+    float GetWidth() const { return static_cast<float>(m_Window->getSize().x); }
+    float GetHeight() const { return static_cast<float>(m_Window->getSize().y); }
 
 #pragma endregion
